improved using Wes' solution
I learnt a little bit about regex and how to use them with the '.match' method for filtering. this meant that I no longer had to use '.toLowerCase()' on the inputted search string and for each item in the array.

added the 'numberWithCommas()' function too, which is a function nabbed from stack overflow (according to Wes') to just add commas within large numbers.