article-less string sorting
this was a concept for me.
using a function and a regex, I was able to sort the band names, and remove any articles (e.g. 'the', 'a', 'an').
using this shallow copy, I could then sort them into a new variable.
sort takes 2 arguments, prev iteration and current iteration. then by comparing each by checking if the current is greater than the previous, it'll sort it higher or lower in the list.
Appending onto a 'ul' element using '.innerHTML' and passing in a '.map' placing each sorted band name into new 'li' elements.