Speech Synthesis
** WebSpeech's SpeechRecognition is no longer supported **

Created a speech sythesiser instead.

first defined a variable storing the window.speechSynthesis interface.

creating a new 'utterance' means to pass text into the interface for it to convert it into speech sythensis.
To do this, created a new SpeechSynthesisUtterance(+ passing through input text here). the utterance has methods for pitch and rate too which I played with too.